Adds handling for start.sh

+++ lamp/scripts/start.sh
@@ -8,48 +8,62 @@
 # Apache.
 ##
 
-rm -f /etc/apache2/sites-enabled/*
-scp /etc/conf/apache/vhost.conf /etc/apache2/sites-available/drupal
-ln -s /etc/apache2/sites-available/drupal /etc/apache2/sites-enabled/drupal
+if [ -f '/etc/conf/apache/vhost.conf' ]; then
+  rm -f /etc/apache2/sites-enabled/*
+  scp /etc/conf/apache/vhost.conf /etc/apache2/sites-available/drupal
+  ln -s /etc/apache2/sites-available/drupal /etc/apache2/sites-enabled/drupal
+fi
 
 ##
 # PHP.
 ##
 
-scp /etc/conf/php/php.ini /etc/php5/apache2/php.ini
-scp /etc/conf/php/php.ini /etc/php5/cli/php.ini
+if [ -f '/etc/conf/php/php.ini' ]; then
+  scp /etc/conf/php/php.ini /etc/php5/apache2/php.ini
+  scp /etc/conf/php/php.ini /etc/php5/cli/php.ini
+fi
 
 ##
 # Mysql.
 ##
 
-scp /etc/conf/mysql/my.cnf /etc/mysql/my.cnf
+if [ -f '/etc/conf/mysql/my.cnf' ]; then
+  scp /etc/conf/mysql/my.cnf /etc/mysql/my.cnf
+fi
 chown -R mysql:mysql /var/lib/mysql
 
 ##
 # Cron.
 ##
 
-scp /etc/conf/cron/* /etc/cron.d/
+if [ -d '/etc/conf/cron' ]; then
+  scp /etc/conf/cron/* /etc/cron.d/
+fi
 
 ##
 # SSHD.
 ##
 
 # Global ssh config.
-scp /etc/conf/sshd/sshd_config /etc/ssh/sshd_config
+if [ -f '/etc/conf/sshd/sshd_config' ]; then
+  scp /etc/conf/sshd/sshd_config /etc/ssh/sshd_config
+fi
 
 # Root user.
-mkdir -p /root/.ssh
-scp /etc/conf/sshd/root_authorized_keys /root/.ssh/authorized_keys
-chmod 400 /root/.ssh/authorized_keys
-chown root:root -R /root/.ssh
+if [ -f '/etc/conf/sshd/root_authorized_keys' ]; then
+  mkdir -p /root/.ssh
+  scp /etc/conf/sshd/root_authorized_keys /root/.ssh/authorized_keys
+  chmod 400 /root/.ssh/authorized_keys
+  chown root:root -R /root/.ssh
+fi
 
 # Deployer user.
-mkdir -p /home/deployer/.ssh
-scp /etc/conf/sshd/deployer_authorized_keys /home/deployer/.ssh/authorized_keys
-chmod 400 /home/deployer/.ssh/authorized_keys
-chown deployer:deployer -R /home/deployer/.ssh
+if [ -f '/etc/conf/sshd/deployer_authorized_keys' ]; then
+  mkdir -p /home/deployer/.ssh
+  scp /etc/conf/sshd/deployer_authorized_keys /home/deployer/.ssh/authorized_keys
+  chmod 400 /home/deployer/.ssh/authorized_keys
+  chown deployer:deployer -R /home/deployer/.ssh
+fi
 
 ##
 # Permissions.
@@ -61,7 +75,9 @@ chown -R deployer:www-data /var/www
 # Rsyslog.
 ##
 
-scp /etc/conf/rsyslog/rsyslog.conf /etc/rsyslog.conf
+if [ -f '/etc/conf/rsyslog/rsyslog.conf' ]; then
+  scp /etc/conf/rsyslog/rsyslog.conf /etc/rsyslog.conf
+fi
 
 ##
 # Supervisord.
