+++ nginx/scripts/start.sh
@@ -12,19 +12,25 @@ chmod -R 775 /etc/nginx/auth
 ##
 
 # Global ssh config.
-scp /etc/conf/sshd/sshd_config /etc/ssh/sshd_config
+if [ -f '/etc/conf/sshd/sshd_config' ]; then
+  scp /etc/conf/sshd/sshd_config /etc/ssh/sshd_config
+fi
 
 # Root user.
-mkdir -p /root/.ssh
-scp /etc/conf/sshd/root_authorized_keys /root/.ssh/authorized_keys
-chmod 400 /root/.ssh/authorized_keys
-chown root:root -R /root/.ssh
+if [ -f '/etc/conf/sshd/root_authorized_keys' ]; then
+  mkdir -p /root/.ssh
+  scp /etc/conf/sshd/root_authorized_keys /root/.ssh/authorized_keys
+  chmod 400 /root/.ssh/authorized_keys
+  chown root:root -R /root/.ssh
+fi
 
 ##
 # Rsyslog.
 ##
 
-scp /etc/conf/rsyslog/rsyslog.conf /etc/rsyslog.conf
+if [ -f '/etc/conf/rsyslog/rsyslog.conf' ]; then
+  scp /etc/conf/rsyslog/rsyslog.conf /etc/rsyslog.conf
+fi
 
 ##
 # Supervisord.
